% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{musicBulkexample}
\alias{musicBulkexample}
\title{example bulk data for this package}
\format{
matrix
}
\source{
Processed RNA-seq data from bulk kidney in healthy and APOL1 mouse (Beckerman et al.)
}
\usage{
data("musicBulkexample")
}
\description{
example bulk data for this package
}
\examples{
data("mouseBrainSubsetSCE")
}
\keyword{datasets}
