% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{musicSCEexample}
\alias{musicSCEexample}
\title{example single cell data for this package}
\format{
SingleCellExperiment
}
\source{
Subset of single cell RNA-seq data of mouse kidney (Park et al.)
}
\usage{
data("musicSCEexample")
}
\description{
example single cell data for this package
}
\examples{
data("musicSCEexample")
}
\keyword{datasets}
