% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMusicResults.R
\name{plotMusicResults}
\alias{plotMusicResults}
\title{Plotting function of runMusic.R}
\usage{
plotMusicResults(
  inSCE,
  analysisType = c("EstCellProp", "PreGroupedClustProp", "SingleCellClust"),
  heatmapTitle = NULL,
  analysisName = NULL,
  useAssay = NULL,
  colDataName = NULL,
  rowDataName = NULL,
  scale = FALSE
)
}
\arguments{
\item{inSCE}{A \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{analysisType}{Character. Specify which function to run
Available options are  "EstCellProp","PreGroupedClustProp","SingleCellClust"}

\item{heatmapTitle}{Character. Title for heatmap; Default is NULL}

\item{analysisName}{Character. User-defined analysis name.
This will be used as the slot name and results can be stored and retrived from SCE object using this name}

\item{useAssay}{Choose the assay to be plotted with the heatmap.}

\item{colDataName}{The column name(s) in \code{colData} that need
to be added to the annotation. See plotSCEHeatmap from SingleCellTK for more info}

\item{rowDataName}{The column name(s) in \code{rowData} that need
to be added to the annotation}

\item{scale}{scale input assay if set to TRUE. Default is FALSE}
}
\value{
SingleCellExperiment object containing the outputs of the
specified algorithms in the \link{colData}
of \code{inSCE}.
}
\description{
A wrapper that plots heatmap and cluster plots for results from runMusic.R
}
\details{
plotMusicResults function
}
