% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMusic.R
\name{reportMusic}
\alias{reportMusic}
\title{Deconvolution of RNASeq data using single cell data}
\usage{
reportMusic(
  inSCE = inSCE,
  bulkData = bulkData,
  analysisName = "NULL",
  analysisType = c("EstCellProp", "PreGroupedClustProp", "SingleCellClust"),
  markers = NULL,
  clusters = "cellType",
  samples = "sampleID",
  preClusterlist = NULL,
  DEmarkers = NULL,
  groups = NULL,
  selectCt = NULL,
  cellSize = NULL,
  ctCov = FALSE,
  verbose = TRUE,
  iterMax = 1000,
  nu = 1e-04,
  eps = 0.01,
  centered = FALSE,
  normalize = FALSE,
  nonZero = TRUE,
  output_file = NULL,
  output_dir = NULL
)
}
\arguments{
\item{inSCE}{A SingleCellExperiment object.}

\item{bulkData}{matrix of bulk RNASeq data}

\item{analysisName}{Character. User-defined analysis name.
This will be used as the slot name and results can be stored and retrived from SCE object using this name}

\item{analysisType}{Character. Specify which function to run
Available options are  "EstCellProp","PreGroupedClustProp","SingleCellClust"}

\item{markers}{List. list of gene names. Same as group.markers option from MuSiC package. The list include differential expressed genes within groups.
List name must be the same as \code{clusters}. Default is NULL}

\item{clusters}{character, the colData of single cell dataset used as clusters; Default is "cellType"}

\item{samples}{. Default is sampleID.
groups = NULL,}

\item{preClusterlist}{list of cell types. The list identify groups of similar cell types.}

\item{DEmarkers}{list of gene names. The list include differential expressed genes within groups. List name must be the same as group.markers.}

\item{groups}{grouping labels, character, the colData of single cell data used as groups}

\item{selectCt}{vector of cell types, default as NULL. If NULL, then use all cell types provided by single cell dataset; NULL, #same as select.ct}

\item{cellSize}{data.frame of cell sizes.same as cell_size; data.frame of cell sizes. 1st column contains the names of cell types, 2nd column has the cell sizes per cell type. Default as NULL. If NULL, then estimate cell size from data;}

\item{ctCov}{logical. If TRUE, use the covariance across cell types; #same as ctCov in MuSiC}

\item{verbose}{logical, default as TRUE.}

\item{iterMax}{numeric, maximum iteration number, same as iter.max in MuSiC. Default 1000}

\item{nu}{regulation parameter, take care of weight when taking reciprocal 1e-04,}

\item{eps}{Threshold of convergence. Default 0.01,}

\item{centered}{logic, subtract avg of Y and D. Default FALSE}

\item{normalize}{logic, divide Y and D by their standard deviation. Default FALSE}

\item{nonZero}{logical, default as TRUE. If true, remove all gene with zero expression, same as non.zero.}

\item{output_file}{name of the output file}

\item{output_dir}{name of output marker}
}
\value{
SingleCellExperiment object containing the outputs of the specified algorithms in the \link{colData} of \code{inSCE}.
}
\description{
A wrapper that performs deconvolution and clustering using MuSiC tool and
SingleCellExperiment object
}
\examples{
\dontrun{
data("musicSCEexample")
data("musicBulkexample")
music_sce <- reportMusic(inSCE = musicSCEexample, 
bulkData = musicBulkexample,
analysisName = "test1",
analysisType = "EstCellProp", 
markers = NULL, 
clusters = "cellType", 
samples = "sampleID", 
preClusterlist = NULL,
DEmarkers = NULL,
groups = NULL, 
selectCt = NULL, 
cellSize = NULL, #same as cell_size 
ctCov = FALSE, #same as ctCov
verbose = TRUE,
iterMax = 1000, #same as iter.max 
nu = 1e-04,
eps = 0.01,
centered = FALSE,
normalize = FALSE,
nonZero = TRUE)
output_file = ,
}
\dontrun{
reportMusicResults(inSCE= music_sce,
                  analysisName = "test2",
                  analysisType = "EstCellProp", 
                  heatmapTitle = "test")
}

}
