% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMusic.R
\name{reportMusicAll}
\alias{reportMusicAll}
\title{reportMusicAll}
\usage{
reportMusicAll(
  inSCE = inSCE,
  bulkData = bulkData,
  analysisName = "NULL",
  analysisType = c("EstCellProp", "PreGroupedClustProp", "SingleCellClust"),
  markers = NULL,
  clusters = "cellType",
  samples = "sampleID",
  preClusterlist = NULL,
  DEmarkers = NULL,
  groups = NULL,
  selectCt = NULL,
  cellSize = NULL,
  ctCov = FALSE,
  verbose = TRUE,
  iterMax = 1000,
  nu = 1e-04,
  eps = 0.01,
  centered = FALSE,
  normalize = FALSE,
  nonZero = TRUE,
  output_file = NULL,
  output_dir = NULL,
  heatmapTitle = NULL
)
}
\description{
reportMusicAll
}
