% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMusic.R
\name{reportMusicResults}
\alias{reportMusicResults}
\title{reportMusicResults}
\usage{
reportMusicResults(
  inSCE,
  analysisName,
  analysisType,
  heatmapTitle,
  output_file = NULL,
  output_dir = NULL
)
}
\description{
reportMusicResults
}
