% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMusic.R
\docType{class}
\name{runMusic}
\alias{runMusic}
\alias{getMusicResults}
\alias{y-class}
\alias{getMusicResults,SingleCellExperiment-method}
\alias{getMusicResults<-}
\title{Deconvolution of RNASeq data using single cell data}
\usage{
runMusic(
  inSCE,
  bulkData,
  analysisName = "NULL",
  analysisType = c("EstCellProp", "PreGroupedClustProp", "SingleCellClust"),
  markers = NULL,
  clusters = "cellType",
  samples = "sampleID",
  preClusterlist = NULL,
  DEmarkers = NULL,
  groups = NULL,
  selectCt = NULL,
  cellSize = NULL,
  ctCov = FALSE,
  verbose = TRUE,
  iterMax = 1000,
  nu = 1e-04,
  eps = 0.01,
  centered = FALSE,
  normalize = FALSE,
  nonZero = TRUE
)

getMusicResults(x, y)

\S4method{getMusicResults}{SingleCellExperiment}(x, y)

getMusicResults(x, y) <- value
}
\arguments{
\item{inSCE}{A \linkS4class{SingleCellExperiment} object with singlecell RNASeq data.}

\item{bulkData}{bulk RNASeq matrix object}

\item{analysisName}{Character. User-defined analysis name.
This will be used as the slot name and results can be stored and retrived from SCE object using this name}

\item{analysisType}{Character. Specify which function to run
Available options are  "EstCellProp","PreGroupedClustProp","SingleCellClust"}

\item{markers}{vector or list of gene names, default as NULL. If NULL, use all genes that provided by both bulk and single cell dataset.}

\item{clusters}{character, the colData of single cell dataset used as clusters; Default is "cellType"}

\item{samples}{Default is sampleID.}

\item{preClusterlist}{list of cell types. The list identify groups of similar cell types.}

\item{DEmarkers}{List. list of gene names. Same as group.markers option from MuSiC package. The list include differential expressed genes within groups.
List name must be the same as \code{clusters}. Default is NULL}

\item{groups}{NULL groups passes the column name of higher-cluster in phenoData.}

\item{selectCt}{vector of cell types, default as NULL. If NULL, then use all cell types provided by single cell dataset; NULL, #same as select.ct}

\item{cellSize}{data.frame of cell sizes.same as cell_size; data.frame of cell sizes. 1st column contains the names of cell types, 2nd column has the cell sizes per cell type. Default as NULL. If NULL, then estimate cell size from data;}

\item{ctCov}{logical. If TRUE, use the covariance across cell types; #same as ctCov in MuSiC}

\item{verbose}{logical, default as TRUE.}

\item{iterMax}{numeric, maximum iteration number, same as iter.max. Default 1000}

\item{nu}{regulation parameter, take care of weight when taking reciprocal 1e-04,}

\item{eps}{Threshold of convergence. Default 0.01,}

\item{centered}{logic, subtract avg of Y and D. Default FALSE,}

\item{normalize}{logic, divide Y and D by their standard deviation. Default FALSE,}

\item{nonZero}{same as non.zero, default is TRUE,}

\item{x}{A \linkS4class{SingleCellExperiment} object with singlecell RNASeq data,}

\item{y}{analysisName,}

\item{value}{A \linkS4class{SingleCellExperiment} object with results stored in metadata}
}
\value{
\linkS4class{SingleCellExperiment} object containing the outputs of the specified algorithms in the \link{colData} of \code{inSCE}.
}
\description{
A wrapper that performs deconvolution and clustering using MuSiC tool and
SingleCellExperiment object
}
\examples{
\dontrun{
data("musicBulkexample")
data("musicSCEexample")
sce <- runMusic(musicSCEexample,musicBulkexample, analysisType = "EstCellProp",analysisName = "test",
analysisType = "EstCellProp", markers = NULL, clusters = "cellType",samples = "sampleID", preClusterlist = NULL,
DEmarkers = NULL,groups = NULL,selectCt = NULL, cellSize = NULL, ctCov = FALSE, verbose = TRUE,iterMax = 1000, 
nu = 1e-04,eps = 0.01,centered = FALSE,normalize = FALSE, nonZero = TRUE)
}

}
